requirement.txt
Django>=1.8.2
Markdown>=2.6.2
PyYAML>=3.11
argparse>=1.2.1
django-filter>=0.10.0
djangorestframework>=3.1.3
mod-wsgi>=4.4.12
wsgiref>=0.1.2
requests>=2.7.0
psycopg2>=2.6
MySQL-python>=1.2.5
